Add an intermediate representation for the new compiler (#419)
This PR is the first draft of the intermediate representation for symbol
resolution, type inference (of option types) and eventually, generation
of `FileDescriptorProto`s.

This PR is essentially just data structure boilerplate: it does not
contain everything (for example, there is no support for services yet)
nor does it include code for constructing from an AST or an imported
FDP. The purpose of this PR is to get the first draft checked in so that
followup PRs can iterate on the API as symbol resolution, type checking,
and option resolution are implemented.