Implement diagnostic deduplication (#450)
This PR implements "diagnostic deduplication", which allows later
diagnostics to suppress earlier diagnostics on the same topic. This is
necessary for cases where a better diagnostic can be emitted in a later
stage of the compiler for something an earlier stage can legalize.

The precise behavior is that if multiple diagnostics have the same
non-empty tag and primary span, only the last diagnostic is kept after
sorting.

I've also added `cmpx` package to make implementing comparison functions
less error-prone, as there are already multiple fairly delicate ones in
the compiler.