Implement most of lowering from IR to FDS (#467)
This PR implements part of the lowering out of the IR into a proto. This
does not implement services, groups, maps, or options; it is intended as
a scaffold that we can use to start writing tests for lowering *to* IR
from AST.

This has no tests; those will be part of the IR tests that will lower
from AST to FDS, which will use this code as the final step. As a
sequencing measure, I've written this code first since the AST->IR tests
will depend on it.