Fix some bugs in the incremental package (#482)
This PR improves the tests and robustness of the `incremental` query
system, by adding checks to catch potential concurrency bugs, and
validation that various error conditions actually do what they're
documented to do.

This PR also implements an optimization for reducing the number of
goroutines that need to be spawned for a particular batch of tasks.