Add IR tests, with a focus on imports (#487)
This PR adds end-to-end tests to the `ir` package. These tests start
from either a single proto file or a collection of them (in a
`proto.yaml` file) and generates diagnostics and an FDS.

This PR also implements `queries.IR`, which is used in the IR tests. It
is not clear to me if coupling this more tightly with the `ir` module
makes sense, but it does mighty feel like it. If we decide to move it
into the `ir` package as `ir.Query`, one thing I would worry about is
that if `Session.Lower` isn't left exposed, there won't be a way to
construct an IR module from user-build AST. Kind of a corner-case, but
one that seems annoying to close off. At the same time, `Session.Lower`
is such a confusing API...

Also, this changes the `ir.Importer` interface to return a general
error. It felt "more correct" for the lowering logic to be where all
errors become diagnostics. I've also made it so that `queries.File` can
be run with or without diagnostics triggering on failure. This is used
by `queries.IR` to delay the diagnostic being created until
`Session.Lower` gets a chance to.