Redo how options are represented in the IR package (#499)
The existing representation for options has proven unworkable for a few
reasons.

1. Syntax options do not map cleanly to FDP options, because `a.b.c`
results in a nested message, so we would need to aggregate options by
prefix.

2. We need to know how to diagnose cases like the following:

```protobuf
option a = {b: c};
option a.d = 42;
```

In the new approach, the `ir.Option` type is gone, and now there is only
`ir.Value`, which is mostly unchanged from before, with the exception
that it will now remember the path node for the option, intended for use
with diagnostics. When lowering something like `option a.b.c`, we will
set the path to the messages implied by each path component as
appropriate, so that diagnostics can say "because of the message value
implicitly set here" or similar.

Each entity that contains options now holds an `arena.Pointer[rawValue]`
corresponding to the `g.p.*Options` message for that entity. This can be
accessed as a `MessageValue` via `blah.Options()`.

Finally, this PR adds previously missing mechanisms for constructing
`ir.Value`s internal to the `ir` package.

This PR does not contain explicit handling for pseudo-options. That will
happen when I implement option name resolution, since it will be clearer
then what the right way to bodge that in is.