Implement symbol table construction (#490)
What it says in the description: build symbol tables for each file. This
finally gets to use the merging code from `slicesx` added many commits
ago.

Each file's symbol table contains every symbol from every file it
transitively imports. This is needed to do global symbol deduplication.
Each `Symbol` can determine whether or not it is visible by consulting a
bitset in `imports`. Each `imports` also now has a map for mapping a
`File` to which index it's in. This is necessary for the symbol table
merging step, which is quite fussy.

Finally, I've added a `symtab.proto` for generating a representation of
the symbol table to include in the test data. This helps ensure that
we're actually populating this information correctly without needing to
perform complex end-to-end tests.