Implement support for synthetic paths (#414)
This PR adds functions to `ast.Nodes` for creating synthetic paths,
which are backed by a synthetic token tree containing `(<the path>)`.

The implementation is somewhat more complicated than I had originally
planned, mostly because of a need to support `Path.Split` on synthetic
paths (panicking in this case felt very rude).