Implement enough evaluation to do field number calculations (#507)
This PR adds an expression evaluator that is currently only used to
evaluate field number expressions. It will be filled out more completely
when the stubs are filled in to implement option evaluation.

There's a lot of things stubbed out in field number evaluation, because
reserved/extension ranges are not implemented yet, and there is a
somewhat complex dependency with options evaluation. In particular,
options can depend on enum values, but enum value validity depends on
options (due to `allow_alias`). Depending on how I choose to break this
dependency cycle, the layout for an `ir.Value` containing an enum value
may instead contain the enum value _index_ instead of a _number_. That
is, `option optimize_for = SPEED` would record the index of SPEED, since
the number is not yet available.