Implement all of constant evaluation (#511)
This PR implements all of the necessary evaluation logic to interpret
all options, including all of the nasty corner cases around `Any`. It
includes extensive diagnostics for various unfortunate corner cases in
what protoc does, and tests to match.

This PR also makes some debugging changes. First, I've added
`report.AnnotateICE`, which can be used to attach span information to
panics. Very helpful for debugging the constant evaluator. Second, the
debug traces of diagnostics shown in test failures are now elided by
default, and can be turned on with a test binary flag.

Finally, I made some changes to how the intern names of language
built-in names are handled, since there will probably be ever-more of
them as we implement more things that depend on options, such as
editions. This code can be found in `intern.go` and `lower.go`.

There may be some validation steps missing in places (UTF-8 strings are
not current validated), but much of it depends on e.g. editions so it's
best left to later.