Fix semantic merge conflict (#530)
There was a semantic merge conflict between #525 (protobuf-go and protoc
update) and #511 (latest change in the experimental/ir package,
for evaluating constant expressions). All that was needed was to
re-generate some golden files that were updated/added in #511. This pull
request does that, but it also adds a make target to make that task
simpler in the future (just `make generate-golden`).