Add a "code DOM" for generating formatted output (#491)
This PR adds a port of my existing Rust library for doing this. This
needed to happen anyways for the formatter we're working on, but I had a
more immediate need of fixing the subtly broken formatting of `yaml.go`,
which is exercised by some of the new IR tests.

This change unfortunately results in churn of all of the existing
checked-in generated YAML, because there were various places where we
did not correctly wrap at 80 columns like it's configured to.