Implement the rest of field number handling (#543)
This PR adds field number, extension range, and reserved range handling
and overlap checking. This is a very complicated process to get right,
because there's many places with potential quadratic behavior, and
generating good diagnostics is tricky.

The strategy I've taken is to implement an "interval intersection set",
which is an interval multimap which allows querying for the intersection
of all intervals which contain a point. This means that we can fill the
map with all of the tag intervals, and then iterate over all of the
entries: disjoint intersections of the elements in the map.

This means, for example, that we can diagnose all overlapping ranges
(treating field numbers as a one-element range), point out which range
caused the overlap, and check that extension numbers match an extension
range, all in O(n log n)[^1] time.

This PR adds extensive tests to check that all sorts of cases are
correctly diagnosed. Eventually, I would like to build "field number
suggestion" based on the data structure added in this PR. It also
contains tests for the interval map.

This PR also includes lowering extension/reserved ranges out of AST and
into IR. It also includes an updated API for working with these that is
more conducive to the API the interval data structure wants to have.

One thing not included is extension number overlap checking. This
requires knowing *all* input files, and then calculating a set of all
extension numbers for every type. That will get implemented later, and
will require a new type of `Opener` (something that lists every
entry-point file).

[^1]: There is one case where we can get quadratic behavior. This is
because inserting to an `interval.Intersect` is actually O(k log n),
where k is the number of intersecting intervals. No valid Protobuf file
can trigger this behavior, but pathological files can, such as the
fragment `reserved 1, 1 to 2, 1 to 3, 1 to 4, ...`, which forces k = n.
This can be mitigated by not processing more ranges after we see N
overlaps (for some constant N). This ensures k = O(N).