Implement options marshaling (#553)
This PR adds code to marshal options into an FDP. There are many
complicated cases; in particular, I have taken care to write marshaling
code that is not accidentally quadratic while dealing with length
prefixes.

Unfortunately, the existing `yaml.go` code cannot deal with unknown
fields well. There really isn't a good option for this, so I chose to
use protoscope to disassemble the unknown fields into something
coherent. This makes it easy to verify that the encoded output is
something reasonably sensible (except for packed fields, which look like
a bit of a mess, but that seems unavoidable).