Implement group support (#551)
This PR adds group syntax support. It is based on #549, since it shares
some of the utilities it adds.

I have added diagnostics for groups. There are some pathological cases
where we seem to get duplicate diagnostics, and I gave up trying to fix
it, because this is a long-deprecated language feature and because the
case where the duplicate diagnostic happens is highly pathological
(multiple groups with the same name, at least one of which does not
start with a capital letter).