Implement RPC services (#554)
This PR implements all of the IR structures necessary to represent
services. It also plumbs all the necessary symbol resolution and options
evaluation steps for services and methods, and is able to lower out of
the IR into ServiceDescriptorProto.

The new types are `ir.Service` and `ir.Method`, which do not have a
concept of a parent type because they're only allowed at the top level
(all nested services are diagnosed by the parser and discarded during
lowering).