Genearlize `Value.AST` and friends to handle values sourced from multiple expressions (#558)
Today, `Value.AST` returns a single node. This does not fully express
how all values could be evaluated. For example, consider the option

```protobuf
option f = { k: [1, 2, 3], x: 5, k: [4, 5] };
```

`f.k` has two disjoint expressions that contribute to it. This PR
enhances `Value` to report *both* of these. `Value.AST` continues to
report the first of these, for ease of writing diagnostics. `Value.ASTs`
returns all of these nodes; `Value.OptionPaths` returns the option paths
that set each of these nodes (which may be distinct, for e.g. `option x
= 1; option x = 2;` for repeated x), and `Value.MessageKeys` returns
either the `FieldExpr.Key` or the corresponding option path for a
contributing expression (convenient for diagnostics).

On top of this, `Elements.AST` now returns the expression that an
element evaluates to. This *does not* line up with the contributing
expression, because a single array can contribute multiple elements. We
use a sparse/dense representation, where `rawValue.elemIndices` is used
to map from an element index to a contributing node index. There is some
subtlety in `lower_eval.go` to build this data structure correctly.