Automate and simplify how builtin symbols are managed in the IR package (#561)
There are many symbols in `descriptor.proto` that get special treatment
in the language. They must always be present, and the compiler ICEs if
they aren't. Previously, resolution logic for these was spread across
three files and required quite a bit of manual work to add a new
builtin.

Now, all of it is collected in `builtins.go`, and the process of
matching up `intern.ID`s to members in `ir.builtins` is automated by
reflection. I've also changed the types of the fields in `ir.builtins`
to make it friendlier to use. This allowed me to delete a lot of
boilerplate!