Make internal descriptor implementations ready for edition 2024 (#550)
This adds implementations of new methods to this module's internal
descriptor representations. These methods implement optional methods
that protobuf-go's `protodesc` package supports to make sure a
descriptor can be round-tripped back to a descriptor proto. These
optional methods are for new edition-2024-specific attributes of
descriptors (for supporting the option imports and symbol visibility
features of edition 2024).

Here's where these optional methods were added to the upstream
descriptor impls and where support for them was added to `protodesc`:
https://go-review.googlesource.com/c/protobuf/+/704415

Keeping this PR a draft until that protobuf-go CL is merged.