Add error handling for Edition 2024 (#548)
This adds error handling to the current parser for Edition 2024. We
expand the list of editions with a known maximum edition value. If the
edition is known we return an error stating the latest supported value.
If the edition is unknown we return a list of support editions. Parsing
is expanded to handle 2024 syntax. Adds parsing of `import option` and
`export` or `local` modifiers. Using the 2024 syntax outside the edition
will error stating the 2024 edition is needed.

Handling of export/local requires special care. Edition 2024 restricts
these keywords. Previous editions and syntax do not. The parser
handles these cases.

---------

Co-authored-by: Josh Humphries <2035234+jhump@users.noreply.github.com>