Skip pathological test cases when race detector is enabled (#565)
This test has recently been failing in CI due to being too
slow, with surprising consistency. In CI it is taking over 20s,
but the onset of this slowness doesn't appear to correlate
to any particular recent commit :(

With race detection and coverage instrumentation disabled,
the tests are *much* faster. So this commit makes it so that
we don't bother running the tests when the race detector
is enabled.