Expand `Syntax` to include all of `google.protobuf.Editions` (#564)
This PR adds Edition 2024, as well as special editions corresponding to
all of the other values in `google.protobuf.Editions`, so we can write
test data that involves them.

From the user's perspective, these extra editions are not real, because
the compiler always generates an error diagnostic. However, because the
compiler will just "keep going", it will emit further diagnostics
consistent with use of this edition.

The "extra" editions are named with a `buf/` prefix, to make it clear
that this is a language extension (even though it's rejected by our
compiler).