Implement Edition 2024 grammar changes in the new compiler (#547)
This PR adds grammar support for Edition 2024. This means we now parse:

- `import option "...";`
- `export message { ... }`
- `export enum { ... }`
- `local message { ... }`
- `local enum { ... }`

We also recognize Edition 2024, but emit an error noting that it is not
fully implemented yet. Diagnostics take into account whether the edition
is 2024 or later when diagnosing the new syntax; if it is older, the
Edition 2024 requirement is cited. If later, it notes that the feature
is not implemented fully yet with an error.

This also changes how parts of the AST are modeled. All definitions,
except for `option`s, can now have modifiers, represented as
`TypePrefixed` values; all the code which determines what a `DeclDef`
now accounts for this. Also, the modifier on a `DeclImport` is now a
list, and we can diagnose repeated modifiers (so people don't get a
weird error when they write `import public option "validate.proto";` or
anything like that).