Simplify toposort state keeping
This simplifies the state required for toposort. A single map of key to
bool is used to track whether or not the value was visited, if it was
set, and if the value has been yieled, if true.

Also a small fix to clear the stack and ensure it is zeroed.