Implement unused import warning, and lowering of source code info extensions (#566)
This PR adds two related features:

1. An unused imports warning, which requires additional book-keeping to
track which direct import caused a transitive import to get imported.
2. Codegen for the Buf source code info extension, which includes unused
imports and whether a `syntax` production exists in the file.

In the process of implementing this, I found a bug in how imports are
built, which this PR fixes. It showed up because import causes weren't
being listed correctly, leading to incorrect diagnosis of unused
imports. I also fixed a bug in `yaml.go` where extensions we're being
printed in the output.