Implement message set constraint validation (#571)
This PR adds verification for the following features of a message set
type:

1. No fields.
2. At least one extension range.
3. Extensions are singular messages.

It also fixes message set number ranges to exclude `0x7fff_ffff`, which
is not part of the valid range.

This also fixes how we look up `message_set_wire_format` to be more
robust (this also applies to `allow_alias`).