Add `report.PageBreak` for breaking up same-file diagnostics (#577)
This PR adds a way to break up cases where way too many snippets appear
in the same file. This is pertinent for diagnostics which want to show
something really far away as context, such as highlighting an edition
keyword.

This works by adding `report.PageBreak`, a diagnostic option that causes
all snippets after it to be placed in a new window, regardless of
whether or not they share a file with the previous snippet.