Add new APIs to `experimental/ir` intended for implementing semantic validation (#575)
The next step for the compiler is to implement the laundry list of
sementic validation the language requires, such as checking that
built-in options are used correctly and things like enums and oneofs are
nonempty.

This PR is split from a much larger working branch that contains all of
this validation; this PR adds the APIs that that branch depends on
without adding validation.

The new APIs include:
1. Several new builtin names.
2. Several new predicates on `ir.Member` and `ir.Type`.
3. Better functions for extracting useful spans out of `ir.Value`, along
with bugfixes to how those spans are populated.
4. A uniform way to find if things are marked as deprecated.
5. New diagnostic types to unify several similar ad-hoc diagnostics.

Several diagnostics have changed because this PR improves them, too.