Add direct imports unconditionally to causes (#583)
Was seeing some false positives in unused imports,
which can happen in the case where a direct import
is also shared with a transitive import. And so, when
using `mapsx.Add` to populate `imports.causes`, if
the transitive import was already added, then the direct
import is not being dropped.