Store query in experimental executor tasks (#593)
This PR moves query information from being passed as a parameter to
being stored directly in the `task` struct. Cycle detection uses the
task query for error reports. This prepares for future work where the
cycle detection will use the tasks from the dependencies. Tasks are also
easier to inspect for debugging. And finally, this fixes a subtle bug on
key eviction. Where we would previously clear the task (`*task =
task{}`) , leading to deleting the set `report.Options`, we now simply
remove the task from the executor's tasks map.