Implement custom message field defaults (#611)
This PR implements the other field pseudo-option, `default`.

To make this work, I've refactored how pseudo-options are built to be
more similar to normal options. Specifically, this is so that the
(invalid) `[default.foo = 1]` syntax works for message-typed fields.

I also added UTF-8 validation of options whose corresponding fields
require UTF-8 validation, according to features. Interestingly,
`default` uniquely does *not* go through this check; in that case, a
warning (pointing out that protoc accepts this by mistake) is issued
instead.

---------

Co-authored-by: Edward McFarlane <3036610+emcfarlane@users.noreply.github.com>