Add `ir.Value.Container` (#624)
This PR adds a function for getting the containing message of an
`ir.Value`, analogous to `ir.Member.Container`.

To ensure the container is correctly set everywhere, I've included
runtime validation of the invariant that adding a value to a message is
only done once, and that setting up the container pointer is always
necessary to insert a value, via the `ir.slot` helper.