Move `report.File` and `report.Span` into `source` (#625)
This PR moves the aforementioned types into the `source` package,
specifically so that `source.Opener` can return them. This is also an
opportunity to decouple the import path of an `ast/ir.File` from
`source.File.Path`, so that the latter can return an arbitrary on-disk
path, rather than the truncated Protobuf import path.