Split the lexer into its own internal package (#629)
This PR splits the `experimental/parser` package such that the lexer is
now a separate package, `experimental/internal/lexer`. The lexer has
also been re-written to make the following configurable:

1. Which keywords are actually lexed as keywords.
2. Which comment and bracket tokens are used.
3. Which string escapes are handled.
4. Details about number literals.

This allows the bulk of the lexer to be re-used by other languages that
we hope to write parsers to.

Additionally, the `experimental/token/keyword` package has been
re-written to make the addition of new keywords easy. There are now
keywords for individual brackets (so, there is `LParen` in addition to
`Parens`), and there are keywords for the comment-delimiting tokens.

In order to make the keyword configuration work, I've written a simple
quadtrie implementation, which is a `map[string]T` that allows for
efficient lookup of longest prefixes.