Implement a new generic expression AST to use with future parsers (#633)
This PR adds `experimental/expr`, a new AST for generic expressions,
which is capable of representing (at minimum) both Protobuf and CEL
expressions, along with many other potential syntaxes.

This is intended to unify our expression parsing framework; eventually,
it should ideally replace the `ast.Expr*` types, but that is not a
short-term goal for this package. For now, it will be used as a testing
ground for a new stand-alone expression language.

This PR also includes an `internal/exprx` package which provides an
expression serialization mechanism for use in tests, similar to
`astx.ToProto`. It also includes new keywords, which do not have
corresponding taxa; that will be rectified as part of the work to
implement the parser.

Finally, I've added `Trie.Prefixes`, which allows iterating over all
prefixes of a string in a trie. This is necessary in the lexer so that
the keyword we select in the lexer does not wind up being the longest
prefix among *all* keywords, not just those the lexer configuration
accepts.