Refactor `keyword` and `internal/lexer` to accomodate more complicated grammars (#634)
This PR contains various interdependent changes to the lexing framework:

1. `token.Punct` is now `token.Keyword`, which indicates a _hard_
keyword, i.e. a keyword that cannot be an identifier in the grammar.
2. `lexer.Lexer` now distinguishes hard and soft keywords.
3. Popular keywords (`for`, `if`, `else`, and operators like `:=`) are
now available in the lexer.
4. Newlines are always their own tokens in the lexer; they can be
promoted to keywords with an impact on the grammar via a configurable
callback, which can then be used to implement newline-terminated
statements (like in Go).
5. Various diagnostics and helpers from `parser` have been factored out
into internal packages for re-use.